backend/pyproject.toml
[tool.poetry]
name = "medical-vision"
version = "0.1.0"
description = ""
authors = ["Ashot Svazyan <ashot.svazyan@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14.1 || >3.14.1,<3.15"
torch = "^2.11.0"
isort = "^8.0.1"
black = "^26.3.1"
pillow = "^12.2.0"
torchvision = "^0.26.0"
kagglehub = "^1.0.0"
tqdm = "^4.67.3"
numpy = "^2.4.4"
openai = "^2.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
